wip: Draft of base common URIBackedMixin
