+++ cpp/arcticdb/storage/s3/nfs_backed_storage.hpp
@@ -16,10 +16,11 @@
 #include <arcticdb/storage/s3/s3_client_accessor.hpp>
 #include <arcticdb/storage/s3/s3_storage.hpp>
 #include <arcticdb/storage/s3/s3_api.hpp>
+#include <arcticdb/storage/s3/uri_backed_storage.hpp>
 
 namespace arcticdb::storage::nfs_backed {
 
-class NfsBackedStorage final : public Storage<NfsBackedStorage> {
+class NfsBackedStorage final : public Storage<NfsBackedStorage>, public arcticdb::storage::uri_backed::URIBackedMixin {
     using Parent = Storage<NfsBackedStorage>;
     friend Parent;
 
@@ -43,23 +44,6 @@ protected:
 
     bool do_key_exists(const VariantKey& key);
 
-    bool do_supports_prefix_matching() {
-        return true;
-    }
-
-    bool do_fast_delete() {
-        return false;
-    }
-
-private:
-    auto& client() { return s3_client_; }
-    const std::string& bucket_name() const { return bucket_name_; }
-    const std::string& root_folder() const { return root_folder_; }
-
-    std::shared_ptr<storage::s3::S3ApiInstance> s3_api_;
-    Aws::S3::S3Client s3_client_;
-    std::string root_folder_;
-    std::string bucket_name_;
 };
 
 inline arcticdb::proto::storage::VariantStorage pack_config(const std::string &bucket_name) {
@@ -86,7 +70,7 @@ inline arcticdb::proto::storage::VariantStorage pack_config(
     return output;
 }
 
-} //namespace arcticdb::nfs_backed
+} // namespace arcticdb::nfs_backed
 
 #define ARCTICDB_NFS_BACKED_STORAGE_H_
 #include <arcticdb/storage/s3/nfs_backed_storage-inl.hpp>
