+++ cpp/arcticdb/storage/s3/uri_backed_storage-inl.hpp
@@ -1,0 +1,16 @@
+/* Copyright 2023 Man Group Operations Limited
+*
+* Use of this software is governed by the Business Source License 1.1 included in the file licenses/BSL.txt.
+*
+* As of the Change Date specified in that file, in accordance with the Business Source License, use of this software will be governed by the Apache License, version 2.0.
+*/
+
+#ifndef ARCTICDB_URI_BACKED_STORAGE_H_
+#error "This should only be included by uri_backed_storage.hpp"
+#endif
+
+#include <arcticdb/storage/s3/uri_backed_storage.hpp>
+
+namespace arcticdb::storage::uri_backed {
+
+} //namespace arcticdb::uri_backed
